Rename: change directory * change profile file directory to profile folder @j1y0ungyoon
